FaceHug logic
Im just trying to get back into the coding stuff and I made it so that the croblin jumps and stays attached to a player for a time determined by the animation event. Then it gets off and the navmesh is activated again. SOUNDS GREAT RIGHT, well when the croblin tries to attack again it drifts right in front of you and starts its hiding phase without attacking.